add portfolio

+++ include/stock.h
@@ -15,20 +15,20 @@ class Stock {
   public:
     virtual ~Stock() = default;
     // Get N prices at time T
-    virtual void getPrices(size_t N, double* out, float u) = 0;
+    virtual void getPrices(size_t N, float* out, float u) = 0;
 };
 
 template <Device d>
 class LogNormalStock : public Stock {
   public:
-    double bias, volatility, start;     
+    float bias, volatility, start;     
     LogNormalStock() = delete;
-    LogNormalStock(double s, float b, float v):
+    LogNormalStock(float s, float b, float v):
      bias(b), volatility(v), start(s) {};
     
-    double getPrice(float u);
+    float getPrice(float u);
     // Get N prices at time u (parallelized)
-    void getPrices(size_t N, double* out, float u) override;
+    void getPrices(size_t N, float* out, float u) override;
 };
 
 }
